(Framework) Refactoring (again)
We take it seriously and combine Symfony and Fatfree feature.
Some component copied (with modification) from Symfony and Fatfree
source code.

Now, each controller should returns an instance of Response.
And we hold each request in a request stack, just like in symfony.

Some of service and parameters registered in default Kernel class.
This framework still lacks of Documentation but we working on it.